+++ playground-tailwind/components/ui/Lightbox.vue
@@ -95,5 +95,5 @@ import {
   LightboxRoot,
   LightboxSlide,
   LightboxViewport,
-} from '@nuxt-photo/nuxt/app'
+} from '@lupinum/nuxt-photo/app'
 </script>